refactor: centralize news API connection with a fetcher, update types news/page.ts
