+++ src/components/index.ts
@@ -54,8 +54,8 @@ export * from './layout/header/Header';
 export * from './layout/header/LanguageSwitcher';
 export * from './layout/header/ListDropdown';
 export * from './layout/header/NavDropdown';
-export * from './layout/header/NavItemRenderer';
 export * from './layout/header/Navigation';
+export * from './layout/header/NavItemRenderer';
 export * from './layout/header/SettingList';
 export * from './layout/header/SpecialToggle';
 export * from './layout/header/UserAvatar';
@@ -100,13 +100,13 @@ export * from './main/map/CustomLayerController';
 export * from './main/map/Map';
 export * from './main/map/MapClicks';
 export * from './main/map/MapSection';
+export * from './main/map/mockTasks';
 export * from './main/map/MultiControlPanel';
 export * from './main/map/PopUpContent';
 export * from './main/map/RadiusWatcher';
 export * from './main/map/ScrollAfterDelay';
 export * from './main/map/StoreMapInstance';
 export * from './main/map/UserLocation';
-export * from './main/map/mockTasks';
 export * from './main/map/config/baseLayerConfig';
 export * from './main/map/filters/AutoZoomOnDistanceFilter';
 export * from './main/map/filters/CategoryFilter';
